benchmark: Reduce number of allmodconfig builds to 5
x86_64 has twelve different toolchains to test, which means a total of
120 allmodconfig builds have to be done. If allmodconfig takes 15
minutes to build on a machine, that means that it will take 30 hours to
run this benchmark, which is pretty lengthy.

Reduce the number of builds for allmodconfig to 5, which should not
lose much fidelity in the result, while halving the amount of time that
a run will take.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>