tc_build: Add basic support for building Rust
There a lot of options missing and it does not do any fancy kind of
build. But it is a start, and it is already useful, e.g. Peter could
have used it to test the new KCFI arity flag that requires LLVM 21 but
upstream Rust still uses LLVM 20.

I took the approach that the new script only takes care of building Rust
provided an existing LLVM, which seemed simple and clear.

Thus add the basic infrastructure, plus a bit of documentation. Add it
to the CI, too.

I successfully built it in a clean Debian 12 and Fedora 41.

Cc: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>